refactor - add WordList component
