+++ src/components/WordItem.jsx
@@ -1,16 +1,16 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
-const WordItem = ({ word }) => {
+const WordItem = props => {
   return (
-    <div className='word-item'>
-      Word: {word}
+    <div>
+      WordItem
     </div>
   )
 }
 
 WordItem.propTypes = {
-  word: PropTypes.string.isRequired
+
 }
 
 export default WordItem
