+++ src/components/WordList.jsx
@@ -1,20 +1,30 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import WordItem from './WordItem'
+import { Draggable } from 'react-beautiful-dnd'
 
-const WordsList = ({ wordsList }) => {
-  return wordsList.map((word, i) => (
-    <div key={i}>
-      <WordItem {...word} />
-    </div>
-  )
+const WordList = ({ wordList, currDragged }) => {
+  return (
+    wordList.map((word, index) => {
+      return (
+        <Draggable draggableId={word.id} index={index} key={word.id}>
+          {(provided) => (
+            <div className="word-item" ref={provided.innerRef} {...provided.draggableProps} {...provided.dragHandleProps} id={index === currDragged ? 'dragging' : 'not-dragging'}>
+              {word.word}
+            </div>
+          )}
+
+        </Draggable>
+      )
+    })
   )
 }
 
-WordsList.propTypes = {
-  wordsList: PropTypes.arrayOf(PropTypes.shape({
+WordList.propTypes = {
+  wordList: PropTypes.arrayOf(PropTypes.shape({
+    id: PropTypes.string.isRequired,
     word: PropTypes.string.isRequired
-  }))
+  })),
+  currDragged: PropTypes.isRequired
 }
 
-export default WordsList
+export default WordList
