+++ src/utils/dragDropUtils.js
@@ -1,9 +1,6 @@
 export const reorderList = (wordList, locationDetails) => {
-  // remove one at index (source index)
-  // destructure array to get the dragged item obj w/ id and word
   const [draggedItem] = wordList.splice(locationDetails.source.index, 1)
 
-  // inserts dragged item at the destination index
   const newOrder = wordList.splice(locationDetails.destination.index, 0, draggedItem)
 
   return newOrder
@@ -14,6 +11,5 @@ export const checkWin = (wordList, winListOrder) => {
     return item.id === winListOrder[index].id
   })
 
-  // check win condition array for a false
   return winCondition.includes(false) ? false : true
 }
